refactor: move definitions to more meaningful places
