+++ src/gestures/Gestures.cpp
@@ -8,27 +8,6 @@
 #include <utility>
 #include <wayfire/touch/touch.hpp>
 
-std::string stringifyDirection(GestureDirection direction) {
-    std::string bind;
-    if (direction & GESTURE_DIRECTION_LEFT) {
-        bind += 'l';
-    }
-
-    if (direction & GESTURE_DIRECTION_RIGHT) {
-        bind += 'r';
-    }
-
-    if (direction & GESTURE_DIRECTION_UP) {
-        bind += 'u';
-    }
-
-    if (direction & GESTURE_DIRECTION_DOWN) {
-        bind += 'd';
-    }
-
-    return bind;
-}
-
 std::string CompletedGesture::to_string() const {
     switch (type) {
         case CompletedGestureType::EDGE_SWIPE:
