+++ src/gestures/Gestures.hpp
@@ -20,17 +20,6 @@ enum class DragGestureType {
     EDGE_SWIPE,
 };
 
-enum TouchGestureDirection {
-    /* Swipe-specific */
-    GESTURE_DIRECTION_LEFT  = (1 << 0),
-    GESTURE_DIRECTION_RIGHT = (1 << 1),
-    GESTURE_DIRECTION_UP    = (1 << 2),
-    GESTURE_DIRECTION_DOWN  = (1 << 3),
-    /* Pinch-specific */
-    // GESTURE_DIRECTION_IN = (1 << 4),
-    // GESTURE_DIRECTION_OUT = (1 << 5),
-};
-
 /**
  * Represents a touch gesture.
  *
