+++ src/gestures/Shared.hpp
@@ -1,5 +1,6 @@
 #pragma once
 #include <cstdint>
+#include <string>
 
 // Swipe params
 constexpr static int EDGE_SWIPE_THRESHOLD              = 10;
@@ -22,3 +23,16 @@ constexpr static uint32_t SEND_CANCEL_EVENT_FINGER_COUNT = 3;
 
 // can be one of @eTouchGestureDirection or a combination of them
 using GestureDirection = uint32_t;
+
+enum TouchGestureDirection {
+    /* Swipe-specific */
+    GESTURE_DIRECTION_LEFT  = (1 << 0),
+    GESTURE_DIRECTION_RIGHT = (1 << 1),
+    GESTURE_DIRECTION_UP    = (1 << 2),
+    GESTURE_DIRECTION_DOWN  = (1 << 3),
+    /* Pinch-specific */
+    // GESTURE_DIRECTION_IN = (1 << 4),
+    // GESTURE_DIRECTION_OUT = (1 << 5),
+};
+
+std::string stringifyDirection(GestureDirection direction);
